+++ aiida/transport/__init__.py
@@ -7,30 +7,6 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
-"""
-Transport module
-"""
-
-import aiida.common
-from aiida.common.exceptions import InternalError
-from aiida.common.extendeddicts import FixedFieldsAttributeDict
-
+"""Transport module."""
 from aiida.transport.transport import Transport
-from aiida.transport.util import TransportFactory
-import aiida.transport.cli
-
-
-def copy_from_remote_to_remote(transportsource, transportdestination, remotesource, remotedestination, **kwargs):
-    """
-    Copy files or folders from a remote computer to another remote computer.
-
-    :param transportsource: transport to be used for the source computer
-    :param transportdestination: transport to be used for the destination computer
-    :param str remotesource: path to the remote source directory / file
-    :param str remotedestination: path to the remote destination directory / file
-    :param kwargs: keyword parameters passed to the final put,
-        except for 'dereference' that is passed to the initial get
-
-    .. note:: it uses the method transportsource.copy_from_remote_to_remote
-    """
-    transportsource.copy_from_remote_to_remote(transportdestination, remotesource, remotedestination, **kwargs)
+from aiida.plugins.factory import TransportFactory
