+++ aiida/transport/cli.py
@@ -109,7 +109,7 @@ def create_option(name, spec):
 
 
 def list_transport_options(transport_type):
-    from aiida.transport.util import TransportFactory
+    from aiida.transport import TransportFactory
     options_list = [create_option(*item) for item in TransportFactory(transport_type).auth_options.items()]
     return options_list
